feat: Switch to using CSS classes for syntax highlighting
I'm not wild about any of the default color schemes, and think
they'll probably clash with my eventual site palette. So, this
commit has the Hugo export of the monokai theme, and the necessary
config to get Chroma to use classes instead of style attributes.
Unfortunately, Hugo modules don't fold the config in nicely (or
at all, as far as I can tell) so I needed to duplicate the config
in the config for my content repo (which is the main module).